Move clowder repo print status into decorator
